grosse maj, le jeu est presque testable a ce stade avec deux couches, pas encore possible de jouer a plusieurs

+++ code_java/jdr2d/src/DAO/PotionDAO.java
@@ -16,7 +16,7 @@ public abstract class PotionDAO {
         effets[3]=rs.getInt("pvmax");
 
         Duration d;
-        d=Duration.of(rs.getLong("duree"), ChronoUnit.SECONDS);
+        d=Duration.of(rs.getInt("duree"), ChronoUnit.SECONDS);
         Potion retour=new Potion(rs.getString("nom_objet"),rs.getInt("poid"),effets,d);
         return retour;
     }
